@@ -1,18 +1,17 @@
 name: kraken_mqtt
 description: kraken MQTT protocol support
-version: 0.0.2
+version: 0.1.0
 author:
-homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  mqtt_client: ^6.0.0
-  kraken: ^0.6.6+2
+  mqtt_client: ^9.3.1
+  kraken: ^0.8.0
 
 dev_dependencies:
   flutter_test:
