@@ -1,22 +1,18 @@
 name: kraken_video_player
 description: Kraken Video Player
-version: 0.3.4
+version: 0.4.0
 homepage: https://openkraken.com
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.2.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  video_player_platform_interface: ^1.0.1
-  meta: "^1.0.5"
-  kraken: ^0.7.3+1
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  video_player_platform_interface: ^4.1.0
+  meta: "^1.3.0"
+  kraken: ^0.8.0
 
 # The following section is specific to Futter.
 flutter:
