@@ -1,17 +1,17 @@
 name: kraken_webview
 description: kraken `<iframe />` tags support.
-version: 0.5.3
+version: 0.6.0
 author: openkraken.com
 homepage: https://openkraken.com/
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  kraken: ^0.7.3+1
+  kraken: ^0.8.0
 
 dev_dependencies:
   flutter_test:
