@@ -20,8 +20,18 @@ public:
   OBJECT_INSTANCE(JSIframeElement)
   JSObjectRef instanceConstructor(JSContextRef ctx, JSObjectRef constructor, size_t argumentCount,
                                   const JSValueRef *arguments, JSValueRef *exception) override;
+protected:
+  JSIframeElement() = delete;
+  explicit JSIframeElement(JSContext *context);
+  ~JSIframeElement();
 
-  class IframeElementInstance : public ElementInstance {
+  static JSValueRef postMessage(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject, size_t argumentCount,
+                                const JSValueRef arguments[], JSValueRef *exception);
+
+  JSFunctionHolder m_postMessage{context, prototypeObject, this, "postMessage", postMessage};
+};
+
+class IframeElementInstance : public ElementInstance {
   public:
     DEFINE_OBJECT_PROPERTY(IFrameElement, 3, width, height, contentWindow)
     DEFINE_PROTOTYPE_OBJECT_PROPERTY(IFrameElement, 1, postMessage)
@@ -32,21 +42,11 @@ public:
     bool setProperty(std::string &name, JSValueRef value, JSValueRef *exception) override;
     void getPropertyNames(JSPropertyNameAccumulatorRef accumulator) override;
 
-    NativeIframeElement *nativeIframeElement;
+    NativeIframeElement *nativeIframeElement{nullptr};
 
   private:
-    double _width;
-    double _height;
-  };
-protected:
-  JSIframeElement() = delete;
-  explicit JSIframeElement(JSContext *context);
-  ~JSIframeElement();
-
-  static JSValueRef postMessage(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject, size_t argumentCount,
-                                const JSValueRef arguments[], JSValueRef *exception);
-
-  JSFunctionHolder m_postMessage{context, prototypeObject, this, "postMessage", postMessage};
+    double _width{0};
+    double _height{0};
 };
 
 using IframePostMessage = void (*)(NativeIframeElement *nativePtr, NativeString *message);
